+++ src/main/java/com/roxon/roulette/model/Result.java
@@ -1,12 +1,14 @@
 package com.roxon.roulette.model;
 
+import java.math.BigDecimal;
+
 public class Result {
   private final Player player;
   private final String bettingNumber;
   private final BetResult betResult;
-  private final double moneyWin;
+  private final BigDecimal moneyWin;
 
-  public Result(Player player, String bettingNumber, BetResult betResult, double moneyWin) {
+  public Result(Player player, String bettingNumber, BetResult betResult, BigDecimal moneyWin) {
     this.player = player;
     this.bettingNumber = bettingNumber;
     this.betResult = betResult;
